feat: migrate describeTable to typescript
